+++ servers/lib/src/files/files.module.ts
@@ -1,16 +1,14 @@
 import { Module } from "@nestjs/common";
-import { ConfigService } from "@nestjs/config";
 import { FilesResolver } from "./files.resolver";
 import { LocalFilesService } from "./services/local-files.service";
 import { GitlabFilesService } from "./services/gitlab-files.service";
-import { DataSourceService } from "./services/data-source.service";
+import { FilesServiceFactory } from "./services/files-service.factory";
 @Module({
   providers: [
     FilesResolver,
     LocalFilesService,
     GitlabFilesService,
-    DataSourceService,
-    ConfigService,
+    FilesServiceFactory,
   ],
 })
 export class FilesModule {}
