+++ servers/lib/src/files/interfaces/files.service.interface.ts
@@ -1,0 +1,5 @@
+// FileService interface
+export interface IFilesService {
+  listDirectory(path: string): Promise<string[]>;
+  readFile(path: string): Promise<string[]>;
+}
