+++ servers/lib/test/integration/files.service.integration.spec.ts
@@ -2,7 +2,7 @@ import { Test, TestingModule } from "@nestjs/testing";
 import { INestApplication } from "@nestjs/common";
 import { FilesModule } from "../../src/files/files.module";
 import { GraphQLModule } from "@nestjs/graphql";
-import { FilesService } from "../../src/files/interfaces/files.service";
+import { FilesService } from "../../src/files/interfaces/files.service.interface";
 import { ConfigService } from "@nestjs/config";
 import * as dotenv from "dotenv";
 import {
