+++ servers/lib/test/unit/files.resolver.unit.spec.ts
@@ -1,5 +1,5 @@
 import { Test, TestingModule } from "@nestjs/testing";
-import { FilesService } from "../../src/files/interfaces/files.service";
+import { FilesService } from "../../src/files/interfaces/files.service.interface";
 import { FilesResolver } from "../../src/files/files.resolver";
 import {
   testDirectory,
